design: [FN-258]설정 폼 디자인 수정
